requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform
azure-functions==1.4.0
torch===1.6.0 -f https://download.pytorch.org/whl/torch_stable.html
torchvision===0.7.0 -f https://download.pytorch.org/whl/torch_stable.html